move datasets to their own dict
