+++ cow_detect/datasets/std.py
